// src/__generated__/seeder.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: b0o/Tq9s47pDVLBSDObNkfaiWsDw5VuVQ0W/FRhc2ilgyn9KGPHqrKnPT7O/xJ6AM/9BzYi8PegmFxppNenpkA==
 */

/* eslint-disable */
// tslint:disable

import Player from './player'

interface Seeder {
  discord_id: bigint
  /**
   * @default nextval('seeder_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'seeder_id'}
  notifications_paused_until: (Date) | null
  /**
   * 0 = 'online', 1 = 'playing', 2 = 'always' 3 = 'never'
   * 
   * @default 1
   */
  notify_when: number
  steam_id: Player['steam_id']
}
export default Seeder;

interface Seeder_InsertParameters {
  discord_id: bigint
  /**
   * @default nextval('seeder_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'seeder_id'}
  notifications_paused_until?: (Date) | null
  /**
   * 0 = 'online', 1 = 'playing', 2 = 'always' 3 = 'never'
   * 
   * @default 1
   */
  notify_when?: number
  steam_id: Player['steam_id']
}
export type {Seeder_InsertParameters}