// src/__generated__/tenant.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: Ay9yzZG6x+40Nf9Ke0RqU4U6wAlHem5gkSSOgvvK2g/+pnX701DjP2nX7t/zVMNom9B2aU4/2ctuSY0OTROmPA==
 */

/* eslint-disable */
// tslint:disable

interface Tenant {
  guild_id: bigint
  /**
   * @default nextval('tenants_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'tenant_id'}
  seed_channel_id: (bigint) | null
  signup_message_id: (bigint) | null
}
export default Tenant;

interface Tenant_InsertParameters {
  guild_id: bigint
  /**
   * @default nextval('tenants_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'tenant_id'}
  seed_channel_id?: (bigint) | null
  signup_message_id?: (bigint) | null
}
export type {Tenant_InsertParameters}